requirements.txt
flask
flask-cors
python-dotenv
opencv-python
numpy
deepface
google-generativeai
gunicorn
